requirements.txt
pyserial
sv_ttk